Fix key-value flyweight::caches not compiling correctly.
Also, for the first time in my life, I've found a valid, safe, and
correct use of a const_cast. After MANY MANY hours of thinking on it
(though this thinking is not reflected by the commit times :P), I've
come to the conclusion that a pointer to a T const, where the T has
already been destructed, is ok to be const_cast away, when the original
pointer to a non-const T was allocated by allocator_traits. You can
check this because an allocator_traits cannot be instantiated with an
allocator that allocates T consts. It is, in fact, undefined. As such, a
const_cast is the only way to then safely deallocate the previously
mutably allocated pointer.

A huge comment on this behavior has been left in the code for others who
may happen to peruse and freak out because 'oh my god, a const_cast was
used'.